Created MBDevice. This class provides information about the device, currently if it is a tablet or phone.
The requestWindowFeature for the indeterminate progress indicator is commented out because on the tablet it kept on running; even though the progress indicator wasn't invoked. Anyway, this should be figured out later.

The MBViewManager is only notified when the MBDialogController explicitly finishes itself, unfortunately. So, implemented the onKeyDown for the back button to finish the activity when its stack is empty.