Changed the MBMetadataService so it can handle a config for smartphones, tablet or both. This is done by invoking setPhoneConfigName, setTabletConfigName or the already existing setConfigName. Also, a method is added to ask for the top level dialog, which is either the dialog name of itself or its parent, a DialogGroup.
The code to build the views for the DialogController has been moved to the view package and has the same structure for the view builders. This means that there is a factory, MBDialogBuilderFactory, to get the appropriate builder. For now, there is a builder for single dialogs and for a split view.

The parsing of the config xml is modified to handle DialogGroups. Keep in mind, that a DialogGroup is nothing more than a Dialog containing other Dialogs. So a DialogGroupDefinition is an extension of DialogDefinition with a list of DialogDefinitions. Because a DialogGroup is a Dialog, the dialog list in MBMvcConfigurationParser contains both groups as single dialogs. However, if a Dialog is part of a DialogGroup, it has a field called parent. That is an indicator that the Dialog is a child and should be handled differently.

At this point there was no possibility to test this code on a 1.6 device. But it is tested on the Xoom and the Nexus One. The test confirms that the config parsed, is based on the device type. It must be said, that the test is performed with the whitelabel core-app.