The getFirstDialog was broken because of the tablet configuration structure. The tablet configuration contains multiple includes for dialogs, pages, outcomes, etc.
An MBConfigurationDefinition has a HashMap of Dialogs. When dialogs.xml is included, a new instance of MBConfigurationDefinition is created. The contents of that new instance is copied in the root configuration. Because the Dialogs are kept in a HashMap, the order is not in the order the dialogs were added. 

To fix all this, when a config is included, not only the dialogs are fetched, but first the firstDialog is checked.

Another thing is that the initial pages were still added to the back stack. So, in the fireInitialOutcomes it is now globally set that during the initial outcomes, the back stack is disabled. This change affected the MBPreparePageInBackgroundRunner.