One shitload of changes:
Fixed the issue for the NetworkOnMainThreadException, introduced in Android 3.0, by doing outcome handling on a separate thread and using a Handler. An outcome is passed to that Handler by sending a Message with an instance of MBOutcome.

MBOutcome has been made Parcelable to be able to put it in a Message. This change had a lot of implications. A lot of definition types had to be made Parcelable too. 

After implementing the MBOutcomeHandler, the MBActivityIndicator started to act weird. So the whole mechanism had to be changed. One positive thing of this change for the MBActivityIndicator, is that the MBViewManager.postShowActivityIndicator is now obsolete.

The changes to the MBActivityIndicator broke the futures and options path. That bug is fixed. But still, on the future selection page, the selection doesn't work. This still needs to be fixed.

Because the new outcome handler is running on a separate thread, there were some problems with displaying the tabs. This is fixed by copying code from the Android Compatibility Package. That is, ActivityCompatHoneycomb and a method in MBViewManager.