Fixed two major issues (one as a result from the other):
The MBActivityIndicator was acting weird because of multithreading issues. This is fixed by doing a check for activity in the MBViewManager.hideActivityIndicator. That method is now synchronized.

Once fixed, we ran into a nasty bug. Namely, after pushing the application to the background, we stopped the MBOutcomeHandlerThread. When reopening the application, we started that thread again. However, when clicking on a button, triggering a handle outcome, it kept displaying the activity indicator. This was caused by the following error: 

Handler{4062bc78} sending message to a Handler on a dead thread

This suggested there was something wrong with the MBOutcomeHandler(Thread). However, the cause of it all laid in the use of AsyncTasks. AsyncTasks must be created and executed on the UI thread. We were doing this from the MBOutcomeHandlerThread. So by doing this on the UI thread, the issue was fixed.