Implemented a new feature: the Toolbar. On Android, it represents the actions of the ActionBar. On iOS, it will represent items for the UIToolBar. The MBToolDefinition needs to be expanded with pre-conditions and maybe more.
The following items are put in the Toolbar: settings, refresh and search. Settings opens a modal dialog with the settings page. Refresh triggers a new RefreshAction. Search is not implemented yet, so displays an error about an empty outcome.

The new RefreshAction does a search for AutoRefreshControllers in the active dialog. For the controllers found, the refresh method is invoked. During that refresh, instead of displaying the "normal" indeterminate progress indicator, the refresh action in the ActionBar is replaced with the indeterminate progress indicator view.

For now, I stole a refresh icon from the Google ioshed application. This icon must be redesigned by Binck.

The tabs in the ActionBar is now a custom view: MBTabBar with MBTabs. The reason for this own implementation is a not fully functional Honeycomb SDK. There were problems removing all tabs, which was necessary for implementing the dropdown view in a tab correctly.

Which brings me to another custom implementation, the MBSpinner. This is our own version of Android's Spinner. It was not possible to set a prompt text/view.

MBRunnable has been moved to the util package, as it was in util.exceptions.