https://dev.itude.com/jira/browse/BINCKAPPS-836
Created the functionality to put the result of an outcome to a relative dialog; that is, left or right. When this is provided, the page will be place left or right in the active dialog group.

Also styled default Android views like EditText, Spinner and Checkbox to display the holo light theme backgrounds. The main theme is black, however, the background of the Binck app is white.