https://dev.itude.com/jira/browse/BINCKAPPS-839
Changed the map of the tools in MBConfigurationDefinition to LinkedHashMap to guarantee the order in which the definitions were added.

Modified the SearchController. Static setters are used to determine which result page must be displayed.

Introduced the method buildInitialView in MBBasicViewController. The logic to determine at which point the views must be built was already in. However, subclasses are generally overriding the onCreateView method, which results into errors/crashes when the fragment is displayed as a dialog. By extracting the building of the views in a separate method, the logic in the MBBasicViewController is now accessible to all subclasses. To use it, subclasses must override buildInitialView, instead of onCreateView/Dialog. This should fix the "requestFeature() must be called before adding content" error.