https://dev.itude.com/jira/browse/BINCKAPPS-952
Changed the Mobbl framework. Specifically the parsing of the resources. First of all, the include parsing functionality has been moved to the, now, abstract MBConfigurationParser. The MBConfigurationDefinition and the MBResourceDefinition now implement the MBIncludableDefinition interface which has the addAll method defined.

Second, the parsing of the resources is now done in the same way as the configuration. That is, the technique was already the same. The structures however were different.

Third, the resource parser is extended, now supporting stated resources to define a resource that is basically a combination of resources already defined in the resources.xml.

Last, MBField, including the definition, is modified so it has an attribute source to link to a resource. Which can now also be a stated resource.

Note: the new pages.xml, resources.xml and statedResources.xml are not committed yet.