https://dev.itude.com/jira/browse/BINCKAPPS-952
Implemented the source attribute that it actually does something. It is implemented for buttons now. The MBResourceService is modified to build the StateListDrawable if the resource is an MBStatedResourceDefinition.

Made the getStatedBackground methods in MBStyleHandler deprecated as those can now be created by using the stated resources.

Made the search result page to refresh on login or logout.