https://dev.itude.com/jira/browse/BINCKAPPS-952
Implemented field of type image button for the order and detail buttons in the search result screen. Lame that another type of button is necessary to build a button without padding you don't want.

Also changed the height of the webview to fill parent. This scales to the size of the web page, which is hyperchill.