https://dev.itude.com/jira/browse/BINCKAPPS-1025 https://dev.itude.com/jira/browse/BINCKAPPS-1017
Refactored the MBTab and MBTabBar a little bit. The responsibility of selecting a tab is now entirely for the tab bar.

The MBRESTServiceDataHandler now handles a timeout differently. That is, it shows a message to the user that there was a timeout instead of showing a message that there is no internet connection.