https://dev.itude.com/jira/browse/BINCKAPPS-1027
Refactored AutoRefreshController. It now posts the new runnable delayed after the refresh is finished. That is, if the new refresh methods are implemented instead of the old refreshInBackground. That method still exists but is made deprecated. RefreshDataInBackground and updateViewsInBackground should be implemented to use the improved functionality.

Another performance issue is fixed. Namely, the RefreshAction was executed twice because the action was defined twice. How nice.