https://dev.itude.com/jira/browse/BINCKAPPS-1026 https://dev.itude.com/jira/browse/BINCKAPPS-1027
Did a lot of stuff. Don't know exactly anymore, but the AutoRefreshController is now finished.

Made an abstract for the indicator classes which makes it easier to have different indicators per device type (smartphone, tablet).

Also removed the handleAllOnWindowActivated after canceling an alert dialog. That is now moved to BinckSession.logOff