https://mobiledev.itude.com/jira/browse/MOBBL-453
Created the builders and the components. The MBBuilderRegistry is
refactored. It has an extra generic so it doesn't only compare strings
but compares any type. This is created to create a relation between a
Class and a Builder. Just decided to do another refactoring to only have
an MBResource with different types. This way it looks more like the
builders for fields and panels.