https://mobiledev.itude.com/jira/browse/MOBBL-775
Fixed pre condition issue regarding child pre conditions within a dialog
group. Because children are also part of the sorted dialogs, the pre
condition of the group evaluates correctly. Because the children don't
have pre conditions, the evaluation is always true. Resulting in the
group being added to the displayed dialogs anyway.

For now, the DialogGroup retains a list of pre conditions to override
the children's pre conditions to avoid them from being displayed.